+++ app/javascript/reactJS/discuss/index.js
@@ -1,0 +1,3 @@
+import "./discuss"
+import "./comments"
+import "./add_discuss"
