+++ app/javascript/reactJS/index.js
@@ -1,1 +1,1 @@
-import "./post" 
+import "./discuss"
