--- app/javascript/reactJS/post/index.js
@@ -1,2 +1,0 @@
-import "./discuss"
-import "./comments"
