+++ app/javascript/styles/index.js
@@ -5,3 +5,4 @@ import "./posts_new.scss"
 import "./single_article.scss"
 import "./user"
 import "./marked.scss"
+import "./add_discuss.scss"
