+++ app/javascript/styles/single_article.scss
@@ -3,7 +3,7 @@
 }
 
 .single-article-wrap{
-  @apply w-6/12 bg-white shadow-md h-full
+  @apply w-7/12 bg-white shadow-md h-full
 }
 
 .single-article-title{
@@ -35,7 +35,7 @@
 }
 
 .single-article-content-wrap{
-  @apply w-4/5 border-b border-gray-200 pt-6 pb-4 flex flex-col items-center
+  @apply w-10/12 border-b border-gray-200 pt-6 pb-4 flex flex-col items-center
 }
 
 .single-article-content-author{
@@ -55,11 +55,15 @@
 }
 
 .single-article-content-body{
-  @apply w-full mt-8 h-full flex flex-col justify-between 
+  @apply w-full mt-8 h-full flex flex-col 
 } 
 
+.single-article-content-body > div:nth-child(1){
+  @apply w-10/12
+}
+
 .single-article-content-body p{
-  @apply flex flex-col py-2 text-gray-600 leading-9
+  @apply flex flex-col py-2 text-gray-600 leading-9 
 }
 
 .single-article-content-span{
@@ -71,7 +75,7 @@
 }
 
 .single-article-reverse-comments{
-  @apply flex justify-start w-4/5 pb-1
+  @apply flex justify-start w-10/12 pb-1
 }
 
 .single-article-reverse-comments > button{
@@ -99,11 +103,11 @@
 }
 
 .single-article-content-input > textarea{
-  @apply  w-4/5 h-24 border-l-2 border-r-2 border-t-2 border-gray-300 rounded-t-md p-3 text-lg resize-none outline-none
+  @apply  w-10/12 h-24 border-l-2 border-r-2 border-t-2 border-gray-300 rounded-t-md p-3 text-lg resize-none outline-none
 }
 
 .single-article-textarea-border{
-  @apply w-4/5 border-2 rounded-b-md border-gray-300 flex justify-end
+  @apply w-10/12 border-2 rounded-b-md border-gray-300 flex justify-end
 }
 
 .single-article-textarea-border > button{
@@ -111,7 +115,7 @@
 }
 
 .single-article-user-comments{
-  @apply m-auto flex flex-col text-base w-4/5 my-10 m-auto
+  @apply m-auto flex flex-col text-base w-10/12 my-10 m-auto
 }
 
 .single-article-user-comments img{
