+++ app/javascript/styles/user/userpage.scss
@@ -1,5 +1,5 @@
 body, html {
-  @apply  h-screen m-0 p-0 relative;
+  @apply  h-screen m-0 p-0 relative break-all;
   
   .user-container{  
     @apply w-full h-screen bg-gray-200;
