+++ app/views/posts/new.html.erb
@@ -1,2 +1,1 @@
-  <h1>新增討論區內容</h1>
-  <%= render "form" %>
+<div id="add-discuss-container" class="add-discuss-container"></div>
